[ chore ] cardFunc 拆分
